requirements.txt
Flask
line-bot-sdk
gunicorn  # 💡 ใช้ Gunicorn ในการรัน Production
google-api-python-client